Make missing test detector reader into a separate module (#10115)
