+++ tools/missing-test-detector/reader/testdata/service/covered_resource_test.go
